Moved KPIComponent to a seperate folder.
