+++ tools/gulptasks/dashboards/api-docs.json
@@ -31,7 +31,8 @@
         "../../../ts/Dashboards/Plugins/DataGridComponent/DataGridComponentOptions.ts",
         "../../../ts/Dashboards/Plugins/HighchartsComponent/HighchartsComponent.ts",
         "../../../ts/Dashboards/Plugins/HighchartsComponent/HighchartsComponentOptions.ts",
-        "../../../ts/Dashboards/Plugins/KPIComponent.ts",
+        "../../../ts/Dashboards/Plugins/KPIComponent/KPIComponent.ts",
+        "../../../ts/Dashboards/Plugins/KPIComponent/KPIComponentOptions.ts",
         "../../../ts/Dashboards/Plugins/NavigatorComponent/NavigatorComponent.ts",
         "../../../ts/Dashboards/Plugins/NavigatorComponent/NavigatorComponentOptions.ts",
         "../../../ts/Data/Connectors/CSVConnectorOptions.d.ts",
